State isn't deleted and re-added if ther'es no change.  Addresses issue #103.

+++ application/classes/Model/Account/State.php
@@ -96,6 +96,12 @@ class Model_Account_State extends Model_Master {
             return false;
         }
         
+        // Get the current state, if it's the same as now we'll ignore this request!
+        $currentState = $account->states->getCurrent();
+        if($currentState->loaded() && $currentState->state == Enum_Account_State::IdToValue($state)){
+            return false;
+        }
+        
         // Remmove all current states.
         foreach($account->states->find_all() as $r){
             $r->delete();
